CategoryList component added
The CategoryList component is designed to display a list of categories on a page. In this commit, importing the necessary dependencies was added and the basic structure of the component was created.
Getting list of categories from Firebase

This commit added a useEffect hook that gets a list of categories from Firebase. Use firebase/firestore/lite's getDocs() function to get a list of documents, then use forEach() to iterate through the documents and add each category's data to a new newCategoryList array using the push() method. The new array is set to the useState variable categoriesList using setCategoriesList().

 Displaying the list of categories

 category listing has been added using map(). For each category, a separate li element is created, which contains a NavLink link to the page with the corresponding category. The category is passed through the slug parameter. The result is displayed inside the ul element in the CategoryList component.