This code contains CSS styles for the Layout component, which is responsible for the overall structure and layout of the site. In particular, properties are set for using CSS Grid in the Layout component. The grid-template-areas defines the areas that will contain the header, aside, main, and footer components. Next, grid-template-columns sets the appropriate columns for these areas. The maximum width for the entire component is set to 890px and the padding from the edges is set.
Separate properties are set for each area, such as grid-area and display. Header gets the display: flex property, which allows you to arrange its content on one line and set the alignment.

aside gets the property grid-area: aside, which indicates the use of the aside area, and main gets the property grid-area: main, which indicates the use of the main area.

The footer gets the property grid-area: footer , indicating that the footer area is being used.