The "CartLink" class defines how the elements inside the block will be placed. "display: flex" sets the box to flex container mode so that the items inside it can be arranged in a row. "justify-content: flex-end" specifies that the elements will be aligned to the right edge of the block.
Next, the "cart-icon" class specifies the maximum size for the cart icon.

The "CartLink a" class defines the styles for the link inside the block. "display: flex" specifies that the link and icon will be inline. "align-items: center" aligns the items vertically. "color: black" specifies the text color of the link, "text-decoration: none" removes the underline from the link, "font-size: 15px" specifies the font size, "font-weight: bold" makes the font bold.

The class "CartLink a:hover" defines the styles for the link when hovering over it. Here the text color is changed to grey.

The code in the "@media screen and (max-width: 600px)" media query defines styles for devices with a screen width less than 600px. Here, the left padding of the "CartLink" block is set to 0 to remove the left padding when the screen shrinks.