the code is responsible for a simple chat using the Socket.IO library and the React framework.
With the useState function, we create two descriptions - an array of messages and the value of the input field of the message inputValue. We then use useEffect to connect to the Socket.IO server and attach a handler to receive messages. We collect receiving messages in an array of its messages using the setMessages function. We also disconnect from the server when checking the component using the return function from useEffect.

The sendMessage function is responsible for sending messages to the server. It uses the socket object created with the message messages to send to the server and clears the value of the input field.

Finally, the returned result is a simple chat that displays messages and has a message input field and a database. When sent, it started in the message array and sent to the server.